Rename duplicate classes
