+++ src/main/kotlin/search/CharacterSearch.kt
@@ -18,7 +18,7 @@ import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
 
-internal object Character {
+internal object CharacterSearch {
     private val lodestoneCssSelectors = Json.parseToJsonElement(
         this::class.java.classLoader.getResource("lodestone-css-selectors/search/character.json") !!
             .readText()
