+++ src/test/java/search/CharacterSearchAsyncTest.java
@@ -1,4 +1,4 @@
-package character;
+package search;
 
 import cloud.drakon.ktlodestone.KtLodestone;
 import cloud.drakon.ktlodestone.search.World;
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-class CharacterAsyncTest {
+class CharacterSearchAsyncTest {
     @Test
     void getCharacterAsync() {
         assertDoesNotThrow(() -> System.out.println(KtLodestone.searchCharacterAsync("Kumokiri Yamitori", World.Cerberus).get()));
